requirements.txt
Django==1.11.6
django-ckeditor==5.3.0
django-js-asset==0.1.1
django-registration-redux==1.8
django-uuslug==1.1.8
itsdangerous==0.24
MySQL-python==1.2.5
python-slugify==1.2.4
pytz==2017.2
Unidecode==0.4.21
